[FAB-13467] Expose more peer keep-alive options
This patchset makes the following peer's keep-alive options
configurable, which is useful to set the config values using core.yaml
or ENV variables.

* peer.keepalive.interval: The duration after which if the server does
not see any activity from the client it pings the client to see if it is
alive;
* peer.keepalive.timeout: The duration the server waits for a response
from the client after sending a ping before closing the connection.

The corresponding sampleconfig/core.yaml and test cases are also update.

FAB-13467 #done

Change-Id: I9cd8343479e8cb10b9b11d181061673cbbc2ff17
Signed-off-by: Baohua Yang <baohua.yang@oracle.com>
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>