FAB-13513 DRY serialization code slightly
A new serialization function to check if a struct has already been
serialized in an opaque way needs to be added.  This CR is a prelude to
the new function and factors some of the repetitive bits into shared
code.  Although the delta is not particularly negative, it will avoid
duplication in the next CR.

Change-Id: Ifa278e323b497e003900c94070af994b4dd77a02
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>