FAB-13512 Add lifecycle integration test
This integration test is supposed to perform basic sanity checking that
the assorted pieces of lifecycle all wire through correctly.

It begins with an empty db, then defines a chaincode for the org,
defines it for the channel, queries the defined namespaces, and finally
queries the defined chaincode (all done through the SCC interface with
mocked db which reads own writes to simulate commitment).

Change-Id: I9cabca1614be7cadc65425e9815ca6e6ed4cebec
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>