fix: sync between levelDB close() & rw APIs()
all Get(), Put(), Write(), NewIterator() in the golevelDB
implementation first checks whether the db is closed. If it
is already closed, it would return an error. When
NewIterator() and Close() rarely interleaves in one of the
unit-test as descirbed in FAB-13374, we get into nil pointer
exception. Though we see this nil pointer exception with
NewIterator() call only (due to the test scenario), we could get
a similar problem with get, put, writeBatch call too. Hence,
instead of using a sync.Mutex, this CR uses a sync.RWMutex
and acquire a write lock during Close() and read lock for all
other functions.

FAB-13374 #done

Change-Id: Ibc6dd67c6c878ed56618215b6a973c650ea9c11b
Signed-off-by: senthil <cendhu@gmail.com>