[FAB-13593] Ledger synchronization in SBE tests
The SBE integration tests include cases where the ledger is updated,
waiting for the event on one peer and then checking the result on
another peer. This patch adds synchronization that waits until both
peers have the same ledger height, i.e. the transaction has been
delivered to both peers.

This is handled similarly in the private data integration tests.

Change-Id: I6b68d3cf9cfeebf6cddd75003b3a946295fb4392
Signed-off-by: Matthias Neugschwandtner <eug@zurich.ibm.com>