[FAB-13640] Create validator routing layer
This change set introduces a routing layer for the validation component. The
existing code performed a number of capability checks directly in the
validation logic, making it very hard to maintain the code. With the new
approach, a routing layer determines - for each block under validation -
which capability is currently in force and calls the appropriate validator
based on that.

Change-Id: Iecccfdd5ae0804bfdf0b89a7f24f1defd4778196
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>