[FAB-10202] remove upgraded chaincode detection
The validator component has logic to detect chaincode upgrade transactions
and invalidate any subsequent invocation of the (older version of the) same
chaincode. However this code is redundant since invocations of any chaincode
currently contain a read dependency on the chaincode entry in the LSCC
namespace, which will ensure that any invocation of a chaincode that follows
a transaction that upgrades the same will be invalidated because of an MVCC
conflict.

Note that it may be possible for a malicious invocation of a chaincode NOT to
introduce said read dependency. However, in fabric's security model, we assume
that the endorsement policy is appropriately set so that such malicious
invocations will be invalidated because of a failure to comply with the
endorsement policy.

Change-Id: I747b0ee7e5bfd7fd864f73b1b5fcc70dbb1c7088
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>