[FAB-13632] decouple vscc validator
The validator component makes use of a vscc validator that encapsulates
endorsement policy checks for endorser transactions. This latter component wil
undergo several changes as part of the lifecycle v2 work, and as such it is
best to decouple it from the validator component proper, clone it and only
apply changes to its v2 instance: older versions may be removed in later
releases simply by removing a directory and updating references.

Change-Id: Ibf328f54bfc7c9bf9e2fcdd998e0673a91f22473
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>