[FAB-12126] Fix discovery client test flake
This change set fixes the following test flake:

Both p15 and p7 belonged to Org D, and had the same height.
The endorsement policy of mycc3 was AND(A,B,C,D) and it was
selecting one peer with the highest height from every org.

Therefore, it can only select either p7 or p15.

The test had a loop in it, that made sure that it selects
both peers with high probability by iterating 10 times over
the selection.

I removed the loop, and made p7 to have a height of p15's height - 1,
for the test to be deterministic.

Also increased the iteration count for a test case by x3, to
make it more stable.

Change-Id: Id4b04197c0fe9229e68aa003e2c9ac6b0566d5d0
Signed-off-by: yacovm <yacovm@il.ibm.com>