FAB-13688 ChaincodeContainerInfo impl
This CR implements the second half of the chaincode.Lifecycle interface,
allowing for the chaincode package to query information about containers
for chaincodes installed in the new lifecycle.

It should work, but it has some obvious performance problems, and does
not correctly enforce the semantics of the new lifecycle with respect to
respecting org definitions.  It is a step in the right direction for the
green-path though it will ultimately need to be rewritten to take
advantage of the impending chaincode cache.

Change-Id: I78cb6ffd28e583fa5390640328d3f2470d51a2a6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>