[FAB-13708] ChaincodeDef lookup func for validation
This CR introduces ChaincodeDefinitionForValidation, a function to be used by
the validation component to retrieve ChaincodeDefinition for a chaincode.
Unlike the existing ChaincodeDefinition function exposed by the same struct,
ChaincodeDefinitionForValidation returns two types of errors so that the
callee can tell apart non-deterministic I/O errors (and halt processing
on the channel) from normal errors (that should lead to the tx being
marked as invalid).

Change-Id: I6fb5b9342936891a897ed6f095f76104759ee220
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>