[FAB-13743] Decouple rwset validator from SB validator
This change set decouples the way the state-based validator examines a
read-write set to determine which policies need to be evaluated. v1.4
requires evaluation of 2 policies (chaincode-level and state-level EP),
whereas 2.0 will also require validation of (implicit) collection policies.
By decoupling this component from the implementation, we can simply
inject a new one in 2.0 to also cater for collection policies.

Change-Id: I86dda56a749be88a67bc1311fee5f582aacc1257
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>