[FAB-13471] lifecycle to handle multiple updates
The ledger calls several HandleChaincodeDeploy for each update
but only a single ChaincodeDeployDone after all invocations
to HandleChaincodeDeploy were made.

The current implementation only supported a single update,
and as a result - a second ChaincodeDeployDone will get stuck
writing to a channel.

This change set makes the Lifecycle in core/cclifecycle
to be able to handle any number of updates in a single block.t

Change-Id: I1d85018af398bd5cb968e42031986a999f6be444
Signed-off-by: yacovm <yacovm@il.ibm.com>