Enhance DeployedChaincodeInfoProvider API
This CR enhances the ImplicitCollections API of
DeployedChaincodeInfoProvider to allow additional params for
chaincodeName and queryexecutor.
This enables the implementation to detect whether the chaincode
is defined in the lscc namespace or in the new +lifecycle namespace
and return the implicit collections only if the chaincode is defined
in the latter.

FAB-13812 #done

Change-Id: I3179b63d37e8346879d37f885658bce248dfcf11
Signed-off-by: manish <manish.sethi@gmail.com>