[FAB-13819] reduce memory consumption in test
The peer/chaincode tests make use of call recording mocks in some of the
tests. In at least one of these tests, an operation on the mock is
called in a tight loop. Since all calls to the mock are recorded, this
results in *massive* memory consumption.

This change removes the recording that resulted in high memory
utilization but does not fix the tight loop. That will be addressed by
FAB-13875.

Change-Id: I7cf3ae3118894684068bedf7ea40ea320f34902f
Signed-off-by: Matthew Sykes <matthew.sykes@gmail.com>
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>