FAB-13876 Wire private data hash hash to cchandler
When adding support for getting private data hashes from the chaincode
shim, the actual wiring into the dispatching function of the handler was
missed.  This is a sign that there are not good tests covering the
handler dispatch, however there is no bvious way to write them and it is
covered in integration tests (how this bug was found).  With that being
said, and the impending changes to larger structure, this should be fine
without UT.

Change-Id: I06460b93279af77b468830b592de4896bcff5a2b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>