FAB-13898 Replace Define/DefineForMyOrg verbs
After feedback from the playback on lifecycle, and additional
discussions with stakeholders, the decision has been made to replace the
verbs DefineChaincode and DefineChaincodeForMyOrg with
CommitChaincodeDefinition and ApproveChaincodeDefinitionForMyOrg.  This
CR does the rote replacement of these.

It additionally removes an old struct "ChaincodeDefinition" which caused
naming conflicts and could be fairly trivially reworked to avoid using.

Change-Id: I51887ab2c2ba8fa5245ff44fcd75901e319ffca0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>