[FAB-13924] Create new policy protobuf
This change set introduces a new protobuf definition for policies used at the
application level. There are current 3 places where such policies are defined
and evaluated: chaincode endorsement policy, state-based endorsement policy
and (soon) collection endorsement policy. Currently we force the usage of
marshalled SignaturePolicyEnvelope structs to express such policies. By
creating this additional struct we lift this restriction and permit for
instance the definition of such policies by reference to the config tree.
This is required for instance by the new chaincode lifecycle, whose system
chaincode requires a policy which is an implicit channel policy.

Change-Id: If69d8f1ddb44524ebea3da526f94716577757beb
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>