[FAB-13580] Fix peer join high-cap-channel panic
Currently, when a old version peer wanna join a newer version channel, the
peer will panic and exit. This behavior is unfriendly as will affect all
existing channels for that peer and the client won't get clear
information about the root cause.

This patchset add a ValidateCapabilities before joining the channel, and
return an error message to the client to indicate the problem is due to
the capabilities requirement. And the peer will still run to process all
existing channels.

Corresponding test cases are implemented to verify with valid and
mismatched capabilities.

Change-Id: I2e0b99ce015331725c5d9e58aa633044507790e2
Signed-off-by: Baohua Yang <baohua.yang@oracle.com>
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>