FAB-13989 Check capabilities for new lifecycle
Although it should not be possible to commit new lifecycle transactions
until after capabilities are turned on because of the existing switches
in the validation path, it will still be confusing to users if they can
invoke these functions without the capability enabled.  This CR protects
the lifecycle CC from invocations on channels without the lifecycle
capability.

Change-Id: I1e877c95ae4507d1b1e626d21a226c419c0822f6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>