FAB-13990 Disable old lifecycle on capability
Once the new lifecycle is enabled, the old lscc namespace becomes read
only and any txes which attempt to write to it will be invalidated.
However, it will be confusing to users if we still allow the old
lifecycle to be invoked for functions which are supposed to commit.
This CR adds a check to ensure that those invocations are rejected.

Change-Id: Ieb6ee1ea2a20e8a14c89ba45b43226dc6b71142d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>