FAB-14015 Specify whether init is required
This CR includes whether init is required as part of a chaincode
specification.  For all chaincodes defined in the old lifecycle, this
will always be true.  For newer chaincodes, the definer may pick.  It is
actually required to cleanly implement the new lifecycle, as system
chaincodes do not require init, and we wish to treat them uniformly with
user chaincodes in the new lifecycle.

Change-Id: I356ddf37472440ea5e4dc0d97c232a621e75a69e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>