FAB-14013 Consolidate chaincode definition fields
The serializer was created to allow for different fields to be
serialized to their own key in the database to avoid the massive
dependency on the entirety of the chaincode definition.  However, as the
information stored in lifecycle has grown, this definition becomes
larger and larger.

We can break the real dependencies down into 3 simple categories:

 1. Endorsement Dependencies -- Things to run the chaincode
 2. Validation Dependencies -- Things to validate the chaincode
 3. Collection Dependencies -- Things required if private data is used

As a compromise for the coarseness of a single proto, and a
simplification to reduce the number of keys in use, we reduce a
chaincode definition to these three parameters.

Change-Id: Ibb6910fc7765d175acb07cf4bce6df734681fa0c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>