[FAB-14005] Create new metakey for state-lev validation
This CR reserves a new string into the state-level metadata map to store
state-based endorsement policies in the new format. The current format forces
policies to be of one specific type. The new format is more flexible and
permits different types of policies to be used seamlessly. However, the new
policy cannot be set using the old metadata key, since its consumer expects a
marshalled SignaturePolicyEnvelope. For this reason the new metadata key is
created. The atomic update semantic of the metadata map ensure that - as
soon as the chaincode uses the new metadata key with the new policy type:
the old metadata key is destroyed (since we don't support metadata upserts).

Change-Id: I752f71dc683629903fbaa063d771b30969f3495d
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>