[FAB-13750] Detect eviction from channel and halt
This change set adds logic in the etcdraft chain that detects
that the node is evicted from the channel, even if it was
disconnected from the cluster while it was evicted.

If a node fails sending to other nodes a consensus message,
it starts suspecting that something is amiss.

It then tries to pull the latest config block, and see
if it is in the channel or not.

If it is not, it:

1) Halts the chain
2) Pulls all blocks until the block that evicts
   the node.

Change-Id: Ic3526d834fbef515119bb899fe62d30fdcf53267
Signed-off-by: yacovm <yacovm@il.ibm.com>