FAB-14035 Add tests for configtxgen encoder
The tests currently covering the configtxgen encoder are almost entirely
based on sample config and really cover almost none of the non-green
paths.  This CR simply replaces all of the existing test with a new
ginkgo suite covering 98.4% of lines.  The uncovered lines are
unreachable and will be removed in a later CR.

Change-Id: I2d2bfd7f4437511258a7832a7df0735fab2a2fbf
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>