FAB-13858 Fix immutable /Channel props on create
Presently, because of an oversight in the way the template channel
configuration is generated on the orderer side, it's not possible to
modify any of the /Channel level policies or values during channel
creation.

This is because the /Channel group has a default Admins policy requiring
that both the /Channel/Application/Admins policy and
/Channel/Orderer/Admins policy be satisfied.  However, the template
configuration never set a /Channel/Application/Admins policy (and
assumed the channel creation tx would set this).

This CR causes the orderer to set the /Channel/Application/Admins if the
incoming channel creation transaction assumes that it exists.

Change-Id: I12ae738adf8605137c19b936fe788668cb98bc91
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>