FAB-12699 Allow orgs to be skipped as foreign
During channel creation transaction generation, it's generally
unnecessary for the organization artifacts like MSP directory and
policies to be present, as these are inherited from the orderer system
channel.  Still, configtxgen throws and error when these cannot be
found.

This CR introduces a new flag which may be attached to an org:
"SkipAsForeign".  This causes configtxgen to ignore the details of the
org beyond its name.  Note, this flag will cause genesis block creation
to fail and appropriate error messages have been added for this case.

Change-Id: I542e8c7fd4269e60c49d21e3416ca25d46525e2d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>