[FAB-14031] Fix flake in etcdraft UT
After reconnect the leader to network, we tick it to resend
previous data. However, if it's ticked too excessively, it
may step down to follower due to `CheckQuorum` being enabled.
This causes the test to fail.

Instead of ticking it, this CR changes one test to simply
enque one more tx to trigger resend of previous data.

Change-Id: I211c7bf59dc6322509336ed8b120d869ea1f42f6
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>