[FAB-14217] Harden etcdraft eviction integration test
The test finds a leader, and then creates a channel,
and then removes the leader from both channels.

However, obviously - that is only the system channel leader,
and it may not be the leader of the application channel.

When it's not the leader - the other nodes might cease communication
with that evicted node (in case it's slow) - and then the test fails
because it doesn't detect its own eviction.

This change set makes the test only evict the leader from the system channel,
and removes the creation of the application channel.

Change-Id: I8ff384c0c2003d137dccf9eba948d25bcc14188e
Signed-off-by: yacovm <yacovm@il.ibm.com>