FAB-14253 Do not mutate etcdraft metadata
Currently, when etcdraft metadata has its 'Marshal' function called,
this causes the structure itself to be modified during the marshal.
Therefore, if the proto is marshaled multiple times, the structure will
fail to marshal the second time.  This CR causes the modificatoins to be
done to a copy instead, then the copy discarded.

Change-Id: I6636ae8cdf57847d43f5151f73019c0c14b1ba50
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>