FAB-14255 Compute anchor peer update less manually
The configtxgen -outputAnchorPeersUpdate is generally a broken mess, as
it can only work against virgin channels which have not otherwise been
modified.  That being said, today, the way this update works, it blows
away any policies which were defined additionally in the org
configurations.  This is a problem as we add more policies to support
things like lifecycle.  This CR simply removes a bunch of manual code
and instead uses the facilities provided by configtxlator to compute a
more sane and automatic update.  In the long term this function should
still be deprecated in favor of using the current channel config as a
baseline rather than simply guessing at versions.

Note -- there are no new tests included in this CR.  This is because the
function will be exercised by the new lifecycle integration tests (how
this was caught), and the existing tests test the other behavior (like
aborting on missing org, etc.).

Change-Id: I423f57a9d064517814a17043194f4d8ffb95345c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>