[FAB-13238] handle container exit during launch
When a chaincode container has successfully started, it is possible for
the chaincode inside the container to fail before registration is
complete. Without this change, the container termination is not observed
by the peer and so, instead of failing quickly, the launch must time
out.

This commit watches for container termination during launch and will
report container exit before registration. This results in better
failure semantics and easier diagnostics.

Change-Id: If3071ee9e406b55a54d4d49e69dac9796e565e5d
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>