[FAB-14275] fix compile_protos.sh
When our build environment was based on ubuntu, we were using GNU
coreutils; now that we've moved to alpine, we're using busybox.

The protobuf compilation script was taking advantage of a GNU extension
to `dirname` by passing multiple path arguments and this does not work
in busybox.

Restructured the script to use xargs to invoke dirname with one path at
a time. Also enabled the pipefail shell option to help catch issues like
this earlier.

Change-Id: Iaf07526be9853a2055925da740aa7bd72113d58f
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>