[FAB-14278] Fix etcdraft flaky UT
UT: lagged node can catch up using snapshot

Even if snapshot interval is 1 Byte, it's NOT guaranteed
that snapshot is taken for EVERY block, because snapshot
is taken async on one go routine (for the exact purpose
of not taking snapshot too excessively, i.e. per every
block)

Although, this is non-deterministic and our assertions
should be more lenient.

Change-Id: Iab9305482d683b00c1331e85df1a6dc78868d06d
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>