FAB-14370 Factor out chaincode parameter Equal
Presently, the only place in the code that compares the equality of the
chaincode parameters for a definition and approval is in the
ApproveChaincodeForOrg function.  However, we need this ability to check
if chaincode parameters match before execution.  This CR factors the
code out into a common function where it can be re-used.

Change-Id: I4453e5aac06e1214777bfaa460620fd792d33e3a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>