FAB-14373 Track interesting key hashes
When we get state updates, they are for the hashed key and have a hashed
value.  We cannot depend on private data being available (as there may
be dissemination errors), so all of our decisions muste be made based on
hashes.  This CR begins to track the 'interesting hashes', ie, the
hashes of the keys which, if modified would alter the approval state of
some chaincode.

Change-Id: I12738875e926632288996540b8915c1c42a7d2f9
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>