FAB-14378 Rewire chaincode lifecycle to cache
Presently, the chaincode lifecycle only uses the public data to retrieve
the chaincode definition, then reads the package filesystem from disk
(for each invocation).  This was a hack to get the green path going, but
critically, it does not confirm that the org has approved the definition
(and of course the performance will be very poor).  This CR re-wires and
re-codes the information retrieval to use the newly added chaincode
cache.

Change-Id: Ia277c399a38337f0cde09b08502267ac2ed7c334
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>