Properly set version for orderer
The orderer/common/metadata provides a Version
variable which is actually never set.  The variable
is provided so that you can actually set the
orderer to a different version than the peer.  However,
this is never actually set at build time so this
change simply sets Version to common/metadata.Version
by default.

FAB-14280 #done

Change-Id: Iaf2efbe1dec298022d7419b3561db22badb5f2bf
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>