FAB-14248 gate migration with Channel V1_4_2
The code in Bundle.ValidateNew() that handles migration should prevent
consensus-type migration if channel capability V1_4_2 is off.

The peer should not process anything under Orderer.ConsensusType - i.e.
Type, Metadata, MigrationState, MigrationContext - exept to verify that
migration commands are not coming in when V1_4_2 is off.

This way, after channel capability V1_4_2 that gates migration is
enabled, the peer should be oblivious to everything under
Orderer.ConsensusType. This way, in the future, if we want to fix or
improve migration, we won't have to upgrade the peers beyond V1_4_2.

Change-Id: I42da0166ef6cd044ec5f35a2d6a56fb1e6b1caa0
Signed-off-by: Yoav Tock <tock@il.ibm.com>