[FAB-14331] - Fix flaky Couch DB health check test
There is a delay between the call to stop container,
and the container getting removed from the docker network.

Modified the test to poll the health check endpoint until
the docker container is completely removed and the health
check returns back the expected result.

Enables the Couch DB health check test [FAB-14333]

Change-Id: If6da40793e63378e6fd79e1d446b66ffeb72af72
Signed-off-by: Saad Karim <skarim@us.ibm.com>