FAB-14014 Remove unneeded serialization logic
Because we are consolidating the chaincode definition fields into a few
different protos, most of the serialization logic around other types
like uint64, and strings is unnecessary.  Therefore, we remove it here
to lower maintenance burden.  We can always go back to before this
commit if we ever wish to recover the logic.

Change-Id: I6993980658ae49c8ebfa39342911571781a1a634
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>