FAB-14407 Remove hash from endorsement info
The chaincode definition really shouldn't reference a hash, because the
existence of a hash depends on a local install package, and not all
nodes running the chaincode (in the long term) will even have a local
package available.  Further, different packagers (say one in Node.js and
one in Java) could produce different hashes even for the same set of
code.

Change-Id: I6f0df1063a4f4fd9f985068217892e0fad97c9cc
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>