[FAB-14501] Handling of CouchDB connection errors
There is a regression caused by https://gerrit.hyperledger.org/r/27558
where the error caused by connection failure to CouchDB does not get handled.

This leads to a nil pointer panic when the nil vdbProvider tries to get
a handle to a database.

This change will handle the CouchDB connection error gracefully and panic
the peer with a good error message (e.g. connection problem,
username/password problem) as was done in prior releases.

Change-Id: I5f459ab92c189de49442aef24fcd80abcd3a0391
Signed-off-by: David Enyeart <enyeart@us.ibm.com>