FAB-8591 Improve channel create/update error
This CR simply wraps the output returned from the configtx processing to
indicate whether the update processing was being treated as channel
creation, or as a channel update.  This is to assist users especially
when they attempt to recreate an extant channel resulting in a
previously difficult to understand error message.

Change-Id: I39d0adc0a4b9b53aa252710575e83309f5fa4524
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>