FAB-14454 Fix broken wal file
If a node crashes while writing to wal, it's likely that the last
wal file is broken, and causes UnexpectedEOF when node is restarted
and tries to load wal.

This CR fixes this by leveraging etcd/wal.Repair util, which essentially
truncates last wal file so it can be properly decoded.

Change-Id: I0bd037578b74e8d51e30ba178b94a39102f3bae5
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>