FAB-14518 Cleanup lifecycle init
After much debate, we have concluded that the best path forward for
chaincode initialization is to have it be handled on the application
side.  This could be through the shims themselves, it could be through
the application toolkit on top, or in the application itself.

However, there is existing legacy chaincode out there which depends on
init, and trying to force the execution of init within the chaincode
shim for v2.0.0 would create an undesirable binding between fabric level
and chaincode build level.

Therefore, this CR adds a new field to the chaincode input proto
indicating whether the invocation is for an invocation of init or not.
In the case that it is, and the chaincode has specified that init is
required, we will follow the traditional init mechanics.

It should be emphasized that this is a _compatibility_ flag and new
chaincodes should move their init logic internally through the standard
invoke.

Change-Id: I80d6f651c7b38fed722909a92c78064f9b89b382
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>