[FAB-14534] add type checking in tests
By declaring unnamed variables of the expected interface, we can ensure
we've implemented all of the required methods of the intended protolator
interface. This activity did catch one messing method on
DynamicConsortiumConfigValue.

Also cleaned up copy pasta in godoc for the protolator interfaces and
switched from assert.Equal to gomega MatchJSON to improve the data
reported during test failures and reduce the likelihood of a test
failure due to changing the order of fields in the JSON.

Change-Id: I404e7cfb05963f50a3053c11e80ea085e3561b00
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>