[FAB-14559] namespace ID from context
In the current code, whenever a chaincode requests an operation (e.g.
GetState, PutState...), the identifier of the ledger namespace is derived
from the internal package ID of the chaincode. This works because currently
the internal package ID of the chaincode must also be identical to the ledger
namespace. To overcome this limitation, this change set derives the identifier
of the ledger namespace from the invocation context. This way, we are now free
to derive the internal package ID independently from the identifier of the
ledger namespace.

Change-Id: I3c0fa2f884c10a593b4577021f7711ca0f14ece8
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>